Move hardError view to a separate file
